support pnpm in upgrade command
