+++ packages/bundler/src/index.ts
@@ -1,4 +1,5 @@
 import {getLatestRemotionVersion} from './get-latest-remotion-version';
+import {getPackageManager} from './get-package-manager';
 import {overrideWebpackConfig} from './override-webpack';
 import {startServer} from './start-server';
 import {cacheExists, clearCache} from './webpack-cache';
@@ -8,8 +9,10 @@ export const BundlerInternals = {
 	cacheExists,
 	clearCache,
 	getLatestRemotionVersion,
+	getPackageManager,
 };
 
 export {bundle} from './bundler';
 export {overrideWebpackConfig};
 export {ProjectInfo} from './project-info';
+export {PackageManager} from './get-package-manager';
