+++ packages/bundler/src/update-available.ts
@@ -1,9 +1,6 @@
-import {existsSync} from 'fs';
-import {join} from 'path';
 import semver from 'semver';
 import {getLatestRemotionVersion} from './get-latest-remotion-version';
-
-type PackageManager = 'npm' | 'yarn' | 'unknown';
+import {getPackageManager, PackageManager} from './get-package-manager';
 
 type Info = {
 	currentVersion: string;
@@ -13,12 +10,6 @@ type Info = {
 	packageManager: PackageManager;
 };
 
-const packageManager = existsSync(join(__dirname, '..', 'yarn.lock'))
-	? 'yarn'
-	: existsSync(join(__dirname, '..', 'package-lock.json'))
-	? 'npm'
-	: 'unknown';
-
 const isUpdateAvailable = async (currentVersion: string): Promise<Info> => {
 	const latest = await getLatestRemotionVersion();
 	return {
@@ -26,7 +17,7 @@ const isUpdateAvailable = async (currentVersion: string): Promise<Info> => {
 		currentVersion,
 		latestVersion: latest,
 		timedOut: false,
-		packageManager,
+		packageManager: getPackageManager(),
 	};
 };
 
@@ -41,7 +32,7 @@ export const isUpdateAvailableWithTimeout = () => {
 				latestVersion: version,
 				updateAvailable: false,
 				timedOut: true,
-				packageManager,
+				packageManager: getPackageManager(),
 			});
 		}, 3000);
 	});
